app/requirements.txt
fastapi
uvicorn
pendulum
python-logging-loki
starlette-prometheus
httpie
